+++ src/graphics.rs
@@ -28,7 +28,7 @@ pub use shader::*;
 pub use texture::*;
 
 use crate::error::Result;
-use crate::math::{FrustumPlanes, Mat4};
+use crate::math::{FrustumPlanes, Mat4, Vec4};
 use crate::platform::{FrontFace, GraphicsDevice, RawIndexBuffer, RawVertexBuffer};
 use crate::window;
 use crate::Context;
@@ -362,6 +362,16 @@ pub fn flush(ctx: &mut Context) {
             &shader.data.handle,
             projection_location.as_ref(),
             ctx.graphics.projection_matrix * ctx.graphics.transform_matrix,
+        );
+
+        let diffuse_location = ctx
+            .device
+            .get_uniform_location(&shader.data.handle, "u_diffuse");
+
+        ctx.device.set_uniform_vec4(
+            &shader.data.handle,
+            diffuse_location.as_ref(),
+            Vec4::new(1.0, 1.0, 1.0, 1.0),
         );
 
         ctx.device.set_vertex_buffer_data(